+++ rpc/cluster/define.go
@@ -1,0 +1,51 @@
+package cluster
+
+import (
+	"fmt"
+
+	"github.com/GZShi/net-agent/rpc/cluster/def"
+	"github.com/GZShi/net-agent/rpc/cluster/service"
+	"github.com/GZShi/net-agent/tunnel"
+)
+
+//
+// NewClient 创建rpc客户端
+//
+func NewClient(t tunnel.Tunnel, ctx tunnel.Context) def.Cluster {
+	return &client{t, ctx}
+}
+
+type client struct {
+	t   tunnel.Tunnel
+	ctx tunnel.Context
+}
+
+//
+// NewService 创建rpc服务
+//
+func NewService() tunnel.Service {
+	return &svc{"cluster", nil, service.New()}
+}
+
+type svc struct {
+	prefix string
+	t      tunnel.Tunnel
+	impl   def.Cluster
+}
+
+func (s *svc) SetAlias(prefix string) {
+	s.prefix = prefix
+}
+
+func (s *svc) Prefix() string {
+	return s.prefix
+}
+
+func (s *svc) Hello(t tunnel.Tunnel) error {
+	s.t = t
+	return nil
+}
+
+func (s *svc) Exec(ctx tunnel.Context) error {
+	return fmt.Errorf("route failed: '%v' not found in '%v'", ctx.GetMethod(), ctx.GetService())
+}
