+++ rpc/dial/def/interface.go
@@ -1,0 +1,5 @@
+package def
+
+type Dial interface {
+	Dial(dialSessionID uint32, network, address string) (connSessionID uint32, err error)
+}
