+++ sample/agent.go
@@ -4,8 +4,8 @@ import (
 	"net"
 
 	"github.com/GZShi/net-agent/exchanger"
-	clusterGen "github.com/GZShi/net-agent/rpc/cluster/gen"
-	dialGen "github.com/GZShi/net-agent/rpc/dial/gen"
+	"github.com/GZShi/net-agent/rpc/cluster"
+	"github.com/GZShi/net-agent/rpc/dial"
 
 	"github.com/GZShi/net-agent/cipherconn"
 	log "github.com/GZShi/net-agent/logger"
@@ -33,7 +33,7 @@ func connectAsAgent(addr, password string) {
 	t := tunnel.New(cc)
 
 	t.Ready(func(t tunnel.Tunnel) {
-		client := clusterGen.NewClient(t, nil)
+		client := cluster.NewClient(t, nil)
 		err := client.Login()
 		if err != nil {
 			log.Get().WithError(err).Error("join cluster failed")
@@ -44,7 +44,7 @@ func connectAsAgent(addr, password string) {
 
 	// agent 默认只支持直接创建连接
 	err = t.BindServices(
-		dialGen.NewService("dial"),
+		dial.NewService(),
 	)
 	if err != nil {
 		log.Get().WithError(err).Error("bind service failed")
